step-1-add-create-react-app
I simply created a new “assets” directory in the theme and inside the directory I ran the command “create-react-app” to create the boilerplate React app inside /assets/app.

At this stage you could change directories into the app directory and run “npm start” and you should be taken to your browser at localhost:3000 with the boilerplate “Create React App” app running.

At this point, you’re viewing the app on a Node server, not in the context of the WordPress theme.

In the next step, we’ll have WordPress load the app assets.