step-1-add-create-react-app
In this step, from the /assets/app directory I ran the command “npm run build" which compiles the Create React App assets into static files that we can use.

I also updated the .gitignore files so they would NOT ignore the files in the build directory, as we want to version them for demo sake.

At this point, we have a static version of the boilerplate Create React App, but we’re still not doing anything with it.

We’ll hook it up to WordPress next!