step-1-add-create-react-app
In this step, we add some functions to our inc/setup.php file to get our theme working with Create React App and setting up some other very basic things, such as adding thumbnail support to the theme and adding thumbnails to the post column in the dashboard.

Create React App stores file names in a manifest.json file, so we’re reading that file and passing the filenames to our enqueue functions to enqueue the build files.

This way, we can run “npm run build” on our app at any time, and WordPress will load those files for us.

At this point, we successfully are running a React Single Page app as our WordPress theme!