step-2-prepare-react-app
In this step, we’re removing the boilerplate that came with Create React App so we can start working on our app.

Here, we remove all the files in the /app/src directory except for the index.js file, which is the main entry point to our app.

And we also modify the index.js file to contain 1 React component, the “App” component, which renders a <div> with an <h1> saying: “WordFlix, your friendly super powered single page app!”

If you run “npm run start” you should see a very minimal app with just your new H1 being rendered.

In the next step, we’ll build the static assets to replace what was built before, and will properly update the WordPress theme to use this new minimal version of our App.