step-3-home-page-styled-components
In this step, we’re creating a separate Homepage component and passing it to our App component.

This is an example of “composability” in React, where you can create components that are composed of multiple components.

In the next step, we’ll look at how we can add some style to the components!