step-3-home-page-styled-components
In this step, we’re importing the stylesheet from Ant Design, which is a fantastic Component Library that we’ll be using in this project. Note that I normally wouldn’t suggest importing the ENTIRE css library, but to get up and running it’s easy. The Ant Design docs have tips on how to only import what you need, which I recommend doing if you choose to use Ant Design Components for any other projects.

Additionally, we’re importing our own stylesheet which imports our fonts to be made available throughout the App.