step-3-home-page-styled-components
In this step, we created a few more styled components to use on our homepage.

You’ll also notice, we imported the “Link” component from ‘react-router’ to use for our button to link to the “/movies” page…I guess that means we should make a Movies page and learn about routing?

First, I think our index file is getting a little big. Before we move to routing, lets put the homepage in it’s own file.