step-3-home-page-styled-components
In this step, we move the HomePage component and the styled components that go along with it to it’s own file under /components/HomePage.js

We also move the necessary “imports” to that file, and at the bottom of the file we export the HomePage as the default export so that other parts of our app can import and use it.

Then, in our index.js file, where we removed all the HomePage and styled components, we now just add HomePage as an import, and we’re at the same spot we were, our files just have some more order to them.

Now we can move on to learn about routing!