step-4-routing
In this step, we add router support using React Router.

In the imports, we import the methods we’ll need from ‘react-router’ and ‘react-router-scroll’.

Then we create 2 new super basic components for “MoviesPage” which will be used to display our list of movies, and our MoviePage, which will be used to display details about an individual movie.

We then have our App component output the Router with history support (so we can use the back button) and add “useScroll” middleware to the router so that when we use the back button it doesn’t scroll to the top, but instead maintains the scroll position.

Now, when we’re on the homepage and click the “Explore Films” button, it will take us to the “/movies” page which will render the “MoviesPage” component and we can then click the link to “Movies Details Page” which will render the “MoviePage” component and allow us to go back to MoviesPage or Home.

Next, we’ll move the MoviesPage and MoviePage components into their own files and style them.