step-5-styling-the-routes
In this step we moved the MoviesPage and MoviePage components into their own files in the “/components” directory.

You’ll notice we had to setup the imports in each of the new component files and export the components, then add the components to our index.js as imports to be used within the router.

In the next step, we’ll use some Ant Design components and create some new Styled Components again to bring some color and design to the page.