step-5-styling-the-routes
In this step, we’re building out the basic layout for the “movies” page, which shows a list of movie posters with their title, and provide links to click into the movie details page.

First, we import styled from “styled-components” and a few layout components from “antd”

Then we build out several styled components for use on the page: MenuBar, Wrapper, PageTitle, MoviesWrapper, and MovieTitle.

Then, we build a MovieCard component that we will repeat to create our grid of movies.

Then, we build out our MoviesPage component to make use of our new styled components, and repeat the columns with the MovieCard to show what our grid of movies will look like.

Soon, we’ll hook up GraphQL queries to the layout to get real data from WordPress populated in the component!

But next, let’s style up our MoviePage.