step-5-styling-the-routes
In this step we create several styled components to build the layout of our page.

First, we import styled from styled-components.

Then, we create several styled components: MenuBar, Wrapper, MovieDetails, Poster and MovieContent

Then we use these new styled components to build out the layout.

Now, we have a styled homepage, movies page and movie detail page!

You might have noticed that some components we’re using in the MoviePage are the same as the MoviesPage, so let’s look at creating a Layout component that we can use to wrap both of these pages!