step-5-styling-the-routes
In this step, we pulled out some repetitive code from our MoviesPage and MoviePage components and created a new PageLayout component.

The PageLayout consisted of the Wrapper, MenuBar and Content, with the Content being dynamic and passing through whatever “children” are passed to it.

Then in the index.js, we create a new Route which uses the PageLayout component, and the MoviePage and MoviesPage routes are nested in it, so that the routes now use the PageLayout as a wrapper and are the components that get passed to it as the children.