step-6-configure-redux-and-apollo
In this step we configure our Redux store.

First, we import all the stuff we need.

Then, we configure our ApolloClient to connect to our GraphQL endpoint at “http://graphql-workshop.dev/graphql” which is the url of our WordPress install’s GraphQL endpoint.

Next, we create our “rootReducer” which is the root object that gets passed to our Redux store. Right now, we’re connecting “apollo” and “routing” to our Redux state tree.

Next, we instantiate the logger which we will use to log our Redux actions, which can be very handy in development.

Finally, we configure the Redux store, passing in our rootReducer, an empty object as our initial state, then we add the following middleware: “thunk” and “loggerMiddleware” as well as Redux Dev Tools support.

This isn’t that much code, but it’s very powerful part of our application.

Next, we’ll look at connecting the Redux store that we’ve just set up to our application.