step-6-configure-redux-and-apollo
In this step we wrap our app with the ApolloProvider, which effectively connects the application to the Redux store we setup.

The first thing we do is import what we need: ApolloProvider from react-apollo, syncHistoryWithStore from react-router-redux, and client (Apollo client) and configureStore from the state/index.js file we created in the previous step.

Then we instantiate our store and the app history.

Then we wrap our app with the ApolloProvider component, passing it our store and our client, then update our router history to use our new history that is synced with our Redux store.

Now, we’ve got an app that’s connected to Redux and has Apollo client setup, which means we’re ready to start making some GraphQL queries and populating our app with some real data!