step-7-graphql-queries
In this step, we write our first GraphQL query for the MoviesPage.

For this page, we mocked up a grid of movies which shows a poster image and the movie title, and links to the movie detail page for the movie, using the movie’s ID in the URL.

So, we need a list of movies, and for each movie we need the id, title and image url.

In WordPress, the database we’re working with is using the “post” post_type to store our movie data, and the posters are set as each post’s featured image. So, we write a GraphQL query to get a list of posts with the data we need.

Then, we connect the query to our component, and export the new connected component.

At this point, if we look in React Dev Tools we can see our connected component has several props that Apollo has provided, specifically the “data” prop, which now contains our posts that we’ve fetched.

In the next step, we’ll use the connected data and render it in our components!