step-7-graphql-queries
In this step, we use the data provided by Apollo (the results of the query) to populate our grid of movies.

Within the MoviesPage component, we first ask for “data” from the component’s props.

Then, we check to see if the data is still loading (a helper convention that Apollo provides to indicate data is still being fetched from the server). If it is loading, we output a div showing things are loading.

If it’s not loading, we output the component, this time instead of our hard-coded mock list of movies, we map through the edges and pass the node to the MovieCard component as the “movie” prop.

Then, in the MovieCard component, we get the “movie” out of the props which were passed to it, and we set the posterUrl value with either a fallback or the sourceUrl of the featuredImage that came back in the query.

Then, we replace our placeholder values in the <img> and <MovieTitle> to output data from our query.

Now, the page is actually populated with data from WordPress via GraphQL!

Next, we’ll look at populating the MoviePage via GraphQL as well.