step-7-graphql-queries
In the MoviesPage component, we set the link to the movie to use the movie.id in the URL.

Then, in the MoviePage component, we use that ID to populate the page. . .let’s look at how.

First, we import gql and grapnel from react-apollos.

Then we write our movieQuery which gets a single post with the id, title, content and featuredImage’ sourceUrl. You’ll also notice that the query makes use of the $id variable.

Below, we connect the graphql query to our component, and this is where we see the variable value being set. We set the “id” variable by getting the “movieId” from the React Router “routeParams”.

Then, we replace our placeholder title, poster url and movie description with the data from our query.

So, now when you click one of the movies from the Movies list page, you’re taken to the movie page with the movieId in the url, and the movie details are now populated for that movie via our GraphQL query.

You might notice though, that we didn’t replace our “Release Date” placeholder. For the sake of this demo, we have the release date stored in postmeta for each film, so we’ll need to include some postmeta in our GraphQL query. In the next step, we’ll look at how to do that!