step-8-graphql-aliasing
In this step we change our GraphQL queries to use some aliases, and update our components to output the data with the aliased field names.

This step doesn’t do anything crucial, it just demonstrates a powerful feature of GraphQL, “Aliasing”. The name of the field that is returned can be set in the query itself, which can make it easy to build declarative, re-usable components and minimize any transformations that are needed on the client.

This probably isn’t the ideal demo of this feature, but wanted to include aliasing just to show it off.