step-9-customize-graphql-schema
Previously we left a placeholder in the MoviePage component for the release date. In this step, we add the releasedDate as a field to the “posts” schema via a filter in our inc/graphql.php file.

Each GraphQL field needs a type, description, and resolve method. In our case, the “releaseDate” will return a string, so we set the type to use the string type already defined by the WPGraphQL plugin.

Then we give a description of the field.

Then, we define a resolve function. The resolve function gets 4 things passed to it: the object being resolved (in our case the WP_Post object, an array of $args on the field, the AppContext which provides details such as who the current user is, etc and the ResolveInfo which provides introspection into where in the query tree the current resolver is, what fields are being queried, and other potentially helpful data). The only piece of context we’re concerned with now is the $post object. We use the $post->ID to look up the “release_date” post meta, then we convert it to a timestamp. We validate whether it’s actually a timestamp, and if so we return a date, otherwise we return null.

Then, in our MoviePage component we add “releaseDate” as a field to our movieQuery, and we replace the “TBA” placeholder text with our movie.releaseDate data that will now be returned in our query.

Next, we’ll look at adding pagination to our MoviesPage!