step-10-graphql-pagination
In this step, we add pagination to our MoviesPage.

First, we create a LoadMore styled component, which will serve as a button that can be clicked to load more movies.

Then, we modify our graphql connection to our MoviesPage component, adjusting what props Apollo will connect to our component. The key part of this is our new “loadMoreEntries” function, which uses the cursor from our last query to fetch more items and merge the results into our existing data set.

Then, we add the loadMore button to be rendered to our page, and add an onClick event so that it will call the “loadMoreEntries” when it’s clicked.

Now we have pagination working!

Next, we’ll look at adding some simple loading indicators for when Apollo is fetching data.