step-11-loading-screens
In this step, we create a new LoadingScreen component that simply renders a loading spinner.

Then we use this new LoadingScreen component in our MoviePage and MoviesPage components to be displayed when Apollo’s “loading” condition is true.

So now, when Apollo is fetching data from our server, the loading indicator shows so that the user knows something is happening.